requirements.txt
fastapi
uvicorn
python-dotenv
firebase-admin
pinecone
openai
google-cloud-firestore
httpx